+++ src/hint_processor/builtin_hint_processor/secp/field_utils.rs
@@ -216,6 +216,7 @@ mod tests {
     };
     use assert_matches::assert_matches;
 
+    use rstest::rstest;
     #[cfg(target_arch = "wasm32")]
     use wasm_bindgen_test::*;
 
@@ -448,10 +449,11 @@ mod tests {
         );
     }
 
-    #[test]
+    #[rstest]
+    #[case(hint_code::IS_ZERO_NONDET)]
+    #[case(hint_code::IS_ZERO_INT)]
     #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
-    fn run_is_zero_nondet_ok_true() {
-        let hint_code = "memory[ap] = to_felt_or_relocatable(x == 0)";
+    fn run_is_zero_nondet_ok_true(#[case] hint_code: &str) {
         let mut vm = vm_with_range_check!();
 
         //Initialize memory
@@ -475,10 +477,11 @@ mod tests {
         check_memory!(vm.segments.memory, ((1, 15), 1));
     }
 
-    #[test]
+    #[rstest]
+    #[case(hint_code::IS_ZERO_NONDET)]
+    #[case(hint_code::IS_ZERO_INT)]
     #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
-    fn run_is_zero_nondet_ok_false() {
-        let hint_code = "memory[ap] = to_felt_or_relocatable(x == 0)";
+    fn run_is_zero_nondet_ok_false(#[case] hint_code: &str) {
         let mut vm = vm_with_range_check!();
 
         //Initialize memory
@@ -502,10 +505,11 @@ mod tests {
         check_memory!(vm.segments.memory, ((1, 15), 0));
     }
 
-    #[test]
+    #[rstest]
+    #[case(hint_code::IS_ZERO_NONDET)]
+    #[case(hint_code::IS_ZERO_INT)]
     #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
-    fn run_is_zero_nondet_scope_error() {
-        let hint_code = "memory[ap] = to_felt_or_relocatable(x == 0)";
+    fn run_is_zero_nondet_scope_error(#[case] hint_code: &str) {
         let mut vm = vm_with_range_check!();
 
         //Initialize memory
@@ -523,10 +527,11 @@ mod tests {
         );
     }
 
-    #[test]
+    #[rstest]
+    #[case(hint_code::IS_ZERO_NONDET)]
+    #[case(hint_code::IS_ZERO_INT)]
     #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
-    fn run_is_zero_nondet_invalid_memory_insert() {
-        let hint_code = "memory[ap] = to_felt_or_relocatable(x == 0)";
+    fn run_is_zero_nondet_invalid_memory_insert(#[case] hint_code: &str) {
         let mut vm = vm_with_range_check!();
 
         //Insert a value in ap before the hint execution, so the hint memory insert fails
@@ -540,18 +545,17 @@ mod tests {
         exec_scopes.assign_or_update_variable("x", any_box!(BigInt::zero()));
         //Execute the hint
         assert_matches!(
-                    run_hint!(vm, HashMap::new(), hint_code, &mut exec_scopes),
-                    Err(HintError::Memory(
-                        MemoryError::InconsistentMemory(
-                            x,
-                            y,
-                            z
-                        )
-                    )) if x ==
-        vm.run_context.get_ap() &&
-                            y == MaybeRelocatable::from(Felt252::new(55i32)) &&
-                            z == MaybeRelocatable::from(Felt252::new(1i32))
-                );
+            run_hint!(vm, HashMap::new(), hint_code, &mut exec_scopes),
+            Err(HintError::Memory(
+                MemoryError::InconsistentMemory(
+                    x,
+                    y,
+                    z
+                )
+            )) if x == vm.run_context.get_ap()
+                && y == MaybeRelocatable::from(Felt252::new(55i32))
+                && z == MaybeRelocatable::from(Felt252::new(1i32))
+        );
     }
 
     #[test]
@@ -560,6 +564,7 @@ mod tests {
         let mut exec_scopes = ExecutionScopes::new();
         let hint_codes = vec![
             hint_code::IS_ZERO_ASSIGN_SCOPE_VARS,
+            // NOTE: this one requires IS_ZERO_ASSIGN_SCOPE_VARS to execute first.
             hint_code::IS_ZERO_ASSIGN_SCOPE_VARS_EXTERNAL_SECP,
         ];
 
@@ -597,7 +602,7 @@ mod tests {
     #[test]
     #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
     fn is_zero_assign_scope_variables_scope_error() {
-        let hint_code = "from starkware.cairo.common.cairo_secp.secp_utils import SECP_P\nfrom starkware.python.math_utils import div_mod\n\nvalue = x_inv = div_mod(1, x, SECP_P)";
+        let hint_code = hint_code::IS_ZERO_ASSIGN_SCOPE_VARS;
         let mut vm = vm_with_range_check!();
         //Skip `x` assignment
         //Execute the hint